requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
av==14.3.0
azure-core==1.33.0
azure-identity==1.21.0
blis==1.3.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cloudevents==1.11.0
cloudpathlib==0.21.0
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
dotenv==0.9.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
exceptiongroup==1.2.2
frozenlist==1.5.0
google-crc32c==1.7.1
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.6.1
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
lazy-object-proxy==1.10.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.4.3
murmurhash==1.0.12
nest-asyncio==1.6.0
numpy==2.2.4
openai==1.73.0
openapi-core==0.19.5
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opentelemetry-api==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-semantic-conventions==0.53b0
packaging==24.2
parse==1.20.2
pathable==0.4.4
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
prance==23.6.21.0
preshed==3.0.9
propcache==0.3.1
pybars4==0.9.13
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
pyee==13.0.0
pyfiglet==1.0.2
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
PyMeta3==0.5.1
pyOpenSSL==25.0.0
pypdfium2==4.30.1
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.24.0
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
scipy==1.15.2
semantic-kernel==1.28.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
thinc==8.3.6
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.1
wrapt==1.17.2
yarl==1.19.0
zipp==3.21.0
